telegram-ai-bot/requirements.txt
python-telegram-bot[job-queue]==21.3
openai>=1.40.0,<2
chromadb>=0.6,<2
aiosqlite==0.20.0
pydantic-settings==2.3.4
